Make the calculation of the build path more automagic
Given the compilation database generally sits at the root of your build
we can automate the process of looking for it. To avoid too much over
calculation I've made the variable local when set.

Additionally I've re-factored the code to extract bits from the JSON so
we can set `default-directory` before the checker is called. This should
ensure the build takes place in exactly the place the database says it
should.

This has removed the warning about the build directory being set but no
compilation database existing. This can be fixed but I suspect a bit
more re-factoring would be required.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>