Expand file name against build directory when comparing filenames
Some build systems (e.g. WAF) build within a separate build
directory (e.g. bin/default) and filenames passed to the compiler are
relative to that directory.  The current code won't match these
filenames.  Instead, expand the file name in the JSON file relative to
the build directory and compare this result to the source filename.

Signed-off-by: Martin Schwenke <martin@meltin.net>